refactor: extract reusable ui components
